requirements.txt
PyYAML==5.4.1
gensim==4.1.2
h5py==3.4.0
numpy==1.21.2
